+++ db/tasks.php
@@ -1,0 +1,32 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+/**
+ * Définition des tâches planifiées
+ * 
+ * 🆕 v1.9.40 : TODO BASSE #6 - Tâche planifiée scan automatique
+ *
+ * @package    local_question_diagnostic
+ * @copyright  2025
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+defined('MOODLE_INTERNAL') || die();
+
+$tasks = [
+    [
+        'classname' => 'local_question_diagnostic\task\scan_broken_links',
+        'blocking' => 0,
+        'minute' => '0',
+        'hour' => '3',
+        'day' => '*',
+        'dayofweek' => '0',  // Dimanche
+        'month' => '*'
+    ]
+];
+